Added local build option

+++ build.sh
@@ -6,11 +6,11 @@ if [[ $? -ne 0 ]]; then
 fi
 ARCHITECTURE="linux/arm/v7,linux/arm64/v8,linux/amd64"
 usage() { 
-    echo "Usage: $0 -t <DOCKER_TAG> [-a <ARCHITECTURE>] [-p Push]"
+    echo "Usage: $0 -t <DOCKER_TAG> [-a <ARCHITECTURE>] [-p Push] [-l Local build]"
     echo "Default architecture: linux/arm/v7,linux/arm64/v8,linux/amd64"
     exit 1 
 }
-while getopts ":a:t:pq" o; do
+while getopts ":a:t:pql" o; do
     case "${o}" in
         t)
             TAG="${OPTARG}"
@@ -23,6 +23,9 @@ while getopts ":a:t:pq" o; do
             ;;
         a)
             ARCHITECTURE="${OPTARG}"
+            ;;
+        l)
+            LOCAL="true"
             ;;
         :)  
             echo "ERROR: Option -$OPTARG requires an argument"
@@ -41,15 +44,19 @@ if [[ "${TAG}" -eq "" ]]; then
     exit 1
 fi
 
-# Setup qemu for multiplatform builds
-docker run --rm --privileged multiarch/qemu-user-static --reset -p yes
-# Create builder for docker
-BUILDER=$(docker buildx create --use)
+if [[ "${LOCAL}" -ne "" ]]; then
+  # Create builder for docker
+  BUILDER=$(docker buildx create --use)
+  # Setup qemu for multiplatform builds
+  docker run --rm --privileged multiarch/qemu-user-static --reset -p yes
+fi
 # Build container on all achitectures in parallel and push to Docker Hub
 eval "docker buildx build $PUSH -t $TAG --platform $ARCHITECTURE . $QUIET"
 # Clean up and return error code for CI system if needed
 ERROR_CODE=$?
-docker buildx rm $BUILDER
+if [[ "${LOCAL}" -ne "" ]]; then
+  docker buildx rm $BUILDER
+fi
 if [[ $ERROR_CODE -ne 0 ]]; then
  exit $ERROR_CODE
 fi
